Add guild validation to routes, update Log model to actually accept guild IDs, document guild and privacy in route responses

+++ api/models.py
@@ -13,7 +13,7 @@ class Log(models.Model):
     type = models.CharField(max_length=40, null=True, help_text='Log type.')
     users = fields.JSONField(default=list, help_text='List of users in log.')
     privacy = models.CharField(max_length=10, default='public')
-    guild = models.IntegerField(null=True)
+    guild = models.BigIntegerField(null=True)
     data = fields.JSONField(default=dict, help_text='Extraneous data.')
 
     class Meta:
